+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/NabHibernateCommonConfig.java
@@ -17,7 +17,6 @@ import ru.hh.nab.hibernate.adapter.NabHibernatePersistenceProvider;
 import ru.hh.nab.hibernate.events.EventListenerRegistryPropagator;
 import ru.hh.nab.hibernate.service.NabServiceContributor;
 import ru.hh.nab.hibernate.service.ServiceSupplier;
-import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceTransactionCallbackFactoryImpl;
 import ru.hh.nab.jdbc.aspect.ExecuteOnDataSourceAspect;
 import ru.hh.nab.jdbc.aspect.ExecuteOnDataSourceTransactionCallbackFactory;
 import ru.hh.nab.jdbc.routing.RoutingDataSourceFactory;
@@ -26,6 +25,7 @@ import ru.hh.nab.jdbc.transaction.TransactionalScope;
 import ru.hh.nab.jdbc.validation.DataSourcesReadyTarget;
 import ru.hh.nab.jdbc.validation.ExecuteOnDataSourceBeanPostProcessor;
 import ru.hh.nab.jpa.NabJpaCommonConfig;
+import ru.hh.nab.jpa.aspect.ExecuteOnDataSourceTransactionCallbackFactoryImpl;
 
 @Configuration
 @Import({
